fixed save data issue and music loop bug

+++ Metroidvania 18 Project/Assets/Scripts/SaveSystem/SaveSystem.cs
@@ -68,6 +68,8 @@ public class SaveSystem
     /// </summary>
     public static void DeleteGameData()
     {
+        if (!File.Exists(_fullPath)) { Debug.LogError("Save System ERROR : Cannot delete file. File does not exists: " + _fullPath); return; }
+
         _fullPath = Application.persistentDataPath + Path.AltDirectorySeparatorChar + _dataFolderName + _saveName;
 
         File.Delete(_fullPath);
